sync

+++ Engine/Source/CoreRHI/Private/RHI/IndexBufferView.cpp
@@ -3,7 +3,7 @@
 namespace CE::RHI
 {
 
-	IndexBufferView::IndexBufferView(RHI::Buffer* buffer, u32 byteOffset, u32 byteCount, IndexFormat indexFormat)
+	IndexBufferView::IndexBufferView(RHI::Buffer* buffer, u64 byteOffset, u64 byteCount, IndexFormat indexFormat)
 		: buffer(buffer)
 		, byteOffset(byteOffset)
 		, byteCount(byteCount)
